Pulled url map converters out into their own file
